+++ Changelog.md
@@ -8,6 +8,8 @@
 * Town folks can now attack monsters lured into town
 * Town folk will pick up weapons dropped by the play to fight monsters and
   return them once the danger is over
+* Hostile monster will now attack the player's friends (i.e. charmed or
+  summoned monsters and townsfolk)
 * Allow throwing of weapons
 
 ### Fixed bugs